Refactor after moving premade samples class
