+++ ocf_data_sampler/torch_datasets/process_and_combine.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 import xarray as xr
-from typing import Optional, Tuple
+from typing import Optional
 
 from ocf_data_sampler.config import Configuration
 from ocf_data_sampler.constants import NWP_MEANS, NWP_STDS
@@ -9,7 +9,6 @@ from ocf_data_sampler.numpy_batch import (
     convert_nwp_to_numpy_batch,
     convert_satellite_to_numpy_batch,
     convert_gsp_to_numpy_batch,
-    convert_site_to_numpy_batch,
     make_sun_position_numpy_batch,
 )
 from ocf_data_sampler.numpy_batch.gsp import GSPBatchKey
@@ -74,15 +73,6 @@ def process_and_combine_datasets(
             }
         )
 
-    if "sites" in dataset_dict:
-
-        da_sites = dataset_dict["sites"]
-        da_sites = da_sites / da_sites.capacity_kwp
-
-        sites_sample = convert_site_to_numpy_batch(da_sites)
-
-        numpy_modalities.append(sites_sample)
-
     if target_key == 'gsp':
         # Make sun coords NumpyBatch
         datetimes = pd.date_range(
@@ -97,56 +87,11 @@ def process_and_combine_datasets(
             make_sun_position_numpy_batch(datetimes, lon, lat, key_prefix=target_key)
         )
 
-    # TODO add sun and time features for site
-
     # Combine all the modalities and fill NaNs
     combined_sample = merge_dicts(numpy_modalities)
     combined_sample = fill_nans_in_arrays(combined_sample)
 
     return combined_sample
-
-
-def process_and_combine_site_sample_dict(
-    dataset_dict: dict,
-    config: Configuration,
-) -> xr.Dataset:
-    """
-    Normalize and combine data into a single xr Dataset
-
-    Args:
-        dataset_dict: dict containing sliced xr DataArrays
-        config: Configuration for the model
-
-    Returns:
-        xr.Dataset: A merged Dataset with nans filled in.
-    
-    """
-
-    data_arrays = []
-
-    if "nwp" in dataset_dict:
-        for nwp_key, da_nwp in dataset_dict["nwp"].items():
-            # Standardise
-            provider = config.input_data.nwp[nwp_key].provider
-            da_nwp = (da_nwp - NWP_MEANS[provider]) / NWP_STDS[provider]
-            data_arrays.append((f"nwp-{provider}", da_nwp))
-          
-    if "sat" in dataset_dict:
-        # TODO add some satellite normalisation
-        da_sat = dataset_dict["sat"]
-        data_arrays.append(("satellite", da_sat))
-
-    if "site" in dataset_dict:
-        # site_config = config.input_data.site
-        da_sites = dataset_dict["site"]
-        da_sites = da_sites / da_sites.capacity_kwp
-        data_arrays.append(("sites", da_sites))
-    
-    combined_sample_dataset = merge_arrays(data_arrays)
-
-    # Fill any nan values
-    return combined_sample_dataset.fillna(0.0)
-
 
 def merge_dicts(list_of_dicts: list[dict]) -> dict:
     """Merge a list of dictionaries into a single dictionary"""
@@ -155,60 +100,6 @@ def merge_dicts(list_of_dicts: list[dict]) -> dict:
     for d in list_of_dicts:
         combined_dict.update(d)
     return combined_dict
-
-def merge_arrays(normalised_data_arrays: list[Tuple[str, xr.DataArray]]) -> xr.Dataset:
-    """
-    Combine a list of DataArrays into a single Dataset with unique naming conventions.
-
-    Args:
-        list_of_arrays: List of tuples where each tuple contains:
-            - A string (key name).
-            - An xarray.DataArray.
-
-    Returns:
-        xr.Dataset: A merged Dataset with uniquely named variables, coordinates, and dimensions.
-    """
-    datasets = []
-
-    for key, data_array in normalised_data_arrays:
-        # Ensure all attributes are strings for consistency
-        data_array = data_array.assign_attrs(
-            {attr_key: str(attr_value) for attr_key, attr_value in data_array.attrs.items()}
-        )
-
-        # Convert DataArray to Dataset with the variable name as the key
-        dataset = data_array.to_dataset(name=key)
-
-        # Prepend key name to all dimension and coordinate names for uniqueness
-        dataset = dataset.rename(
-            {dim: f"{key}__{dim}" for dim in dataset.dims if dim not in dataset.coords}
-        )
-        dataset = dataset.rename(
-            {coord: f"{key}__{coord}" for coord in dataset.coords}
-        )
-
-        # Handle concatenation dimension if applicable
-        concat_dim = (
-            f"{key}__target_time_utc" if f"{key}__target_time_utc" in dataset.coords
-            else f"{key}__time_utc"
-        )
-
-        if f"{key}__init_time_utc" in dataset.coords:
-            init_coord = f"{key}__init_time_utc"
-            if dataset[init_coord].ndim == 0:  # Check if scalar
-                expanded_init_times = [dataset[init_coord].values] * len(dataset[concat_dim])
-                dataset = dataset.assign_coords({init_coord: (concat_dim, expanded_init_times)})
-
-        datasets.append(dataset)
-
-    # Ensure all datasets are valid xarray.Dataset objects
-    for ds in datasets:
-        assert isinstance(ds, xr.Dataset), f"Object is not an xr.Dataset: {type(ds)}"
-
-    # Merge all prepared datasets
-    combined_dataset = xr.merge(datasets)
-
-    return combined_dataset
 
 def fill_nans_in_arrays(batch: dict) -> dict:
     """Fills all NaN values in each np.ndarray in the batch dictionary with zeros.
